Generate a (U,W) mixture graph

Generate a (U,W) mixture graph from a dense graphon W and a mass partition corresponding to a line graph graphon U.

Usage

sample_mixed_graph(W, wts, nd, ns, p, option)

Arguments

W

The dense graphon. This is a symmetric matrix with values in [0,1]

wts

The degree proportions of the hub degrees. Need to add up to 1. This is the mass partition corresponding to the line graph graphon U.

nd

The number of nodes in the dense part of the graph

ns

The number of nodes in the sparse part of the graph

p

The nodes to be added as a proportion of the edges in the dense part

option

Graph joining option. If option == 1 then a disjoit union is considered. If option == 2 the two graphs are joined randomly with the number of edges specified by p.

Value

A graph sampled from the (U,W) mixture.

Examples

library(igraph)
W <- matrix(0.1, nrow = 100, ncol = 100)
wts <- c(0.5, 0.3, 0.2)
ns <- 200
nd <- 100
p <- 0.5
gr <- sample_mixed_graph(W, wts, nd, ns, p, option = 2)
gr
#> IGRAPH 87edd3f U--- 300 936 -- Star
#> + attr: name_1 (g/c), name_2 (g/c), mode_1 (g/c), mode_2 (g/c),
#> | center_1 (g/n), center_2 (g/n), name (g/c), mode (g/c), center (g/n)
#> + edges from 87edd3f:
#>  [1] 1--  5 1-- 16 1-- 28 1-- 38 1-- 39 1-- 42 1-- 64 1-- 69 1-- 70 1-- 81
#> [11] 1-- 97 1--100 1--246 2--  8 2-- 31 2-- 33 2-- 65 2-- 66 2-- 74 2--114
#> [21] 2--144 2--147 2--256 3-- 30 3-- 66 3-- 72 3-- 78 3-- 85 3--100 3--212
#> [31] 4--  7 4-- 11 4-- 16 4-- 17 4-- 21 4-- 28 4-- 34 4-- 51 4-- 56 4-- 76
#> [41] 4-- 81 4-- 89 4-- 93 5-- 15 5-- 50 5-- 67 5-- 70 5-- 73 5-- 79 5-- 81
#> [51] 5-- 87 5-- 97 6-- 15 6-- 16 6-- 31 6-- 47 6-- 52 6-- 55 6-- 56 6-- 68
#> [61] 6-- 77 6-- 88 6-- 91 6-- 98 6--194 6--238 6--270 7--  9 7-- 10 7-- 11
#> + ... omitted several edges