Generate a (U,W) mixture graph

Generate a (U,W) mixture graph from a dense graphon W and a mass partition corresponding to a line graph graphon U.

Usage

sample_mixed_graph(W, wts, nd, ns, p = 0.5, option = 2)

Arguments

W

The dense graphon. This is a symmetric matrix with values in [0,1]

wts

The degree proportions of the hub degrees. Need to add up to 1. This is the mass partition corresponding to the line graph graphon U.

nd

The number of nodes in the dense part of the graph

ns

The number of nodes in the sparse part of the graph

p

The nodes to be added as a proportion of the edges in the dense part

option

Graph joining option. If option == 1 then a disjoit union is considered. If option == 2 the two graphs are joined randomly with the number of edges specified by p.

Value

A graph sampled from the (U,W) mixture.

Examples

library(igraph)
W <- matrix(0.1, nrow = 100, ncol = 100)
wts <- c(0.5, 0.3, 0.2)
ns <- 200
nd <- 100
p <- 0.5
gr <- sample_mixed_graph(W, wts, nd, ns, p, option = 2)
gr
#> IGRAPH 920dae1 U--- 300 996 -- Star
#> + attr: name_1 (g/c), name_2 (g/c), mode_1 (g/c), mode_2 (g/c),
#> | center_1 (g/n), center_2 (g/n), name (g/c), mode (g/c), center (g/n)
#> + edges from 920dae1:
#>  [1] 1-- 12 1-- 16 1-- 18 1-- 28 1-- 38 1-- 54 1-- 58 1-- 63 1-- 64 1-- 65
#> [11] 1-- 86 2--  6 2--  8 2-- 23 2-- 34 2-- 45 2-- 47 2-- 55 2-- 65 2-- 68
#> [21] 2-- 71 2-- 86 2-- 96 3--  6 3-- 18 3-- 28 3-- 36 3-- 40 3-- 57 3-- 59
#> [31] 3-- 71 3-- 74 3-- 79 3-- 84 3-- 87 3-- 94 3--177 4-- 20 4-- 45 4-- 49
#> [41] 4-- 55 4-- 57 4-- 59 4-- 60 4-- 67 4-- 77 4-- 82 4-- 90 4-- 91 4-- 98
#> [51] 4-- 99 4--247 4--298 5--  8 5-- 21 5-- 27 5-- 48 5-- 49 5-- 68 5-- 80
#> [61] 5-- 82 5-- 90 5-- 95 5--212 6-- 14 6-- 16 6-- 23 6-- 31 6-- 36 6-- 38
#> + ... omitted several edges